backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
psycopg2-binary==2.9.9
django-filter==23.5
drf-yasg==1.21.7
